OHRM5X-2335: Fixed errors in Leave module
Problem 1: The getLeaveRequestsPaginator in LeaveRequestDao.php was causing an error due to needing to be grouped by ohrm_leave_request.id. Due to the left join of ohrm_leave, it is not possible to determine how to group by the ID since each ohrm_leave_request record has multiple ohrm_leave records.

Fix: I decided to reduce the number of ohrm_leave records by taking MIN(ohrm_leave.date) which results in only one ohrm_leave record per ohrm_leave_request record. Additionally I selected only ohrm_leave_request.id but the other fields can be taken as well since they are functionally dependent. As a result of this change, I had to change the sort field in LeaveRequestSearchFilterParams.php to match the alias of MIN(ohrm_leave.date).

Problem 2: Functions in LeaveRequestDao.php and LeaveEntitlementDao.php that are used in EmployeeLeaveBalanceAPI were causing issues due to the distinct operator being used with order by.

Fix: The fix was to simply use the alias of the IDENTITY(ohrm_leave_entitlement.id) in the ORDER BY clause. Since ohrm_leave_entitlement.id was directly given it was causing an error since that field was not being used in DISTINCT.