OHRM5X-2335: Fixed getTimesheetItemsPaginatorForEmployeeReport
Changes in EmployeeReportData.php, TimesheetDao.php, and TimesheetDaoTest.php. I changed getTimesheetItemsPaginatorForEmployeeReport to select only the required elements (project name, activity name, customer name and total duration of activities). Now the function does not return timesheet items but instead a multidimensional array with the selected elements.

As a result, I had to make minor changes to the tests. Instead of accessing the TimesheetItem object, the value can be taken directly from the array.